requirements.txt
redis
gevent
requests
urllib3
pycrypto
Pillow
flask
flask-script>=2.0.5
PyMySQL>=0.8
flask_sqlalchemy
flask-debugtoolbar
Flask-WTF
Flask-Bcrypt
Flask-Security
Flask-Restful
flask-babelex
flask_migrate
WTForms-Components
werkzeug
waitress
flasgger
xlwt
webassets
# flask-babel
# flask-bootstrap
# pymongo==3.5.1
# Flask-MongoEngine>=0.9.5
# flask-login==0.3.2
# flask-uploads
# Flask-Session
# Celery
# Flask-Celery-Helper
# Flask-Cache
